Починил Рантаймы: ретиалите мобов, proc/rot, а также увеличил тик на Dripping Cum с Секунды на Пять Секунд.

+++ code/modules/holiday/halloween/halloween.dm
@@ -193,12 +193,12 @@
 	unsuitable_atmos_damage = 0
 	var/timer
 
-/mob/living/simple_animal/hostile/retaliate/clown/insane/clown/Initialize(mapload)
+/mob/living/simple_animal/hostile/retaliate/clown/insane/Initialize(mapload)
 	. = ..()
 	timer = rand(5,15)
 
-//mob/living/simple_animal/hostile/retaliate/clown/insane/Retaliate()
-	//return
+/mob/living/simple_animal/hostile/retaliate/clown/insane/Retaliate()
+	return
 
 /mob/living/simple_animal/hostile/retaliate/clown/insane/ex_act()
 	return
@@ -206,14 +206,16 @@
 /mob/living/simple_animal/hostile/retaliate/clown/insane/wave_ex_act(power, datum/wave_explosion/explosion, dir)
 	return power
 
-/mob/living/simple_animal/hostile/retaliate/clown/insane/Life()
+/mob/living/simple_animal/hostile/retaliate/clown/insane/FindTarget()
+	. = ..()
 	timer--
 	if(target)
 		stalk()
-	return
 
 /mob/living/simple_animal/hostile/retaliate/clown/insane/proc/stalk()
 	var/mob/living/M = target
+	if(!M)
+		return
 	if(M.stat == DEAD)
 		playsound(M.loc, 'sound/spookoween/insane_low_laugh.ogg', 500, 1)
 		qdel(src)
@@ -227,6 +229,7 @@
 	stalk(target)
 
 /mob/living/simple_animal/hostile/retaliate/clown/insane/AttackingTarget()
+	FindTarget()
 	return
 
 /mob/living/simple_animal/hostile/retaliate/clown/insane/adjustHealth()
